feat: TextAreaField 컴포넌트 분리, AnswerBox로 컴포넌트명 변경
