+++ src/components/common/input/AnswerBox.tsx
@@ -1,0 +1,48 @@
+import { ChangeEvent, forwardRef, TextareaHTMLAttributes, useState } from 'react';
+
+import Label from '../Label';
+import TextAreaField from './TextAreaField';
+
+interface AnswerBoxProps extends TextareaHTMLAttributes<HTMLTextAreaElement> {
+  labelText: string;
+}
+
+const AnswerBox = forwardRef<HTMLTextAreaElement, AnswerBoxProps>(
+  ({ labelText, maxLength, disabled, required, placeholder, onChange, ...props }, ref) => {
+    const [text, setText] = useState('');
+
+    const handleChange = (e: ChangeEvent<HTMLTextAreaElement>) => {
+      setText(e.target.value);
+
+      if (onChange) onChange(e);
+    };
+
+    return (
+      <div className='gap-2xs flex w-[35rem] flex-col'>
+        <Label
+          hierarchy='normal'
+          weight='normal'
+          isRequired={required}
+          textColor={`${disabled ? 'text-object-assistive-dark' : 'text-object-neutral-dark'}`}
+        >
+          {labelText}
+        </Label>
+        <TextAreaField
+          {...props}
+          ref={ref}
+          value={text}
+          placeholder={placeholder}
+          onChange={handleChange}
+          maxLength={maxLength}
+          disabled={disabled}
+          required={required}
+        />
+        <div
+          className={`${disabled ? 'text-object-disabled-dark' : 'text-object-assistive-dark'} peer-focus:text-object-neutral-dark body-sm cursor-default self-end`}
+        >{`${text.length}/${maxLength || 0}`}</div>
+      </div>
+    );
+  },
+);
+
+export default AnswerBox;
